Split legacy BNCI datasets by year
